refactor(list): names to lowercase
ISSUES CLOSED: #81
