+++ list/src/button/index.tsx
@@ -1,1 +1,1 @@
-export * from './Button'
+export * from './button'
