+++ list/src/condition/index.tsx
@@ -1,1 +1,1 @@
-export * from './Condition'
+export * from './condition'
