+++ list/src/icons/trashIcon.tsx
