+++ list/src/layout/index.tsx
@@ -1,4 +1,4 @@
-export * from './Layout'
-export * from './Box'
-export * from './Row'
-export * from './Column'
+export * from './layout'
+export * from './box'
+export * from './row'
+export * from './column'
